Game modes renamed to scenarios
