+++ src/scenarios/Scenario.java
@@ -1,10 +1,10 @@
-package gameModes;
+package scenarios;
 
 /**
  * Created by Aunmag on 2016.11.22.
  */
 
-public abstract class GameMode {
+public abstract class Scenario {
 
     public abstract void update();
